refactor(text): move Text class into its own namespace
