+++ YDotNet/Document/Doc.cs
@@ -1,8 +1,8 @@
 using YDotNet.Document.Events;
 using YDotNet.Document.Options;
 using YDotNet.Document.Transactions;
-using YDotNet.Document.Types;
 using YDotNet.Document.Types.Maps;
+using YDotNet.Document.Types.Texts;
 using YDotNet.Infrastructure;
 using YDotNet.Native.Document;
 using YDotNet.Native.Document.Events;
